avoid double promotions

+++ src/examples/ekf_att_pos_estimator/estimator_utilities.cpp
@@ -76,7 +76,7 @@ void swap_var(float &d1, float &d2);
 
 float Vector3f::length() const
 {
-    return sqrt(x*x + y*y + z*z);
+    return sqrtf(x*x + y*y + z*z);
 }
 
 void Vector3f::zero()
